@@ -1,14 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <DebugType>Full</DebugType>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Flurl.Http" Version="2.4.2" />
-    <PackageReference Include="MailKit" Version="2.1.3" />
-    <PackageReference Include="OneSignal.RestAPIv3.Client" Version="1.0.3" />
+    <PackageReference Include="MailKit" Version="2.5.1" />
+    <PackageReference Include="OneSignal.RestAPIv3.Client" Version="1.1.1" />
   </ItemGroup>
 
   <ItemGroup>
