@@ -3,13 +3,11 @@ using Sharebook.Jobs;
 
 namespace ShareBook.Api.Controllers.Generic
 {
-    public class BaseJobController: Controller
+    public class BaseJobController: ControllerBase
     {
         protected IJobExecutor _executor;
         protected string _validToken;
 
         protected bool _IsValidJobToken() => (Request.Headers["Authorization"].ToString() == _validToken);
     }
-
-    
 }
