@@ -10,10 +10,10 @@ namespace ShareBook.Api.Controllers
 {
     [Route("api/[controller]")]
     [EnableCors("AllowAllHeaders")]
-    public class JobController: BaseJobController
+    public class JobController : BaseJobController
     {
-
-        public JobController(IJobExecutor executor, IOptions<ServerSettings> settings)
+        public JobController(IJobExecutor executor,
+                             IOptions<ServerSettings> settings)
         {
             _executor = executor;
             _validToken = settings.Value.JobExecutorToken;
@@ -22,9 +22,12 @@ namespace ShareBook.Api.Controllers
         // TODO: voltar 300 no Throttle
         [HttpGet("Executor")]
         [Throttle(Name = "JobExecutor", Seconds = 5, VaryByIp = false)]
-        public IActionResult Executor(){
-            if (!_IsValidJobToken()) return Unauthorized();
-            else                     return Ok(_executor.Execute());
+        public IActionResult Executor()
+        {
+            if (!_IsValidJobToken())
+                return Unauthorized();
+
+            return Ok(_executor.Execute());
         }
 
     }
