@@ -4,10 +4,11 @@ namespace ShareBook.Api.ViewModels
 {
     public class LoginUserVM
     {
-        [Required]
+        [Required(ErrorMessage = "Email é obrigatório")]
+        [EmailAddress(ErrorMessage = "Email no formato incorreto")]
         public string Email { get; set; }
 
-        [Required]
+        [Required(ErrorMessage = "Senha obrigatório")]
         public string Password { get; set; }
     }
 }
