@@ -8,11 +8,11 @@ namespace ShareBook.Domain.Validators
         #region Messages
         public const string Email = "O email é obrigatório";
         public const string EmailFormat = "O formato do email está inválido";
-        public const string Password = "A senha é obrigatória";
+        public const string Password = "A senha é obrigatória.";
         public const string Name = "O nome é obrigatório";
-        public const string PostalCode = "O cep é obrigatório";
-        public const string PostalCodeInvalid = "O formato do cep está inválido";
-        public const string Linkedin = "O seu endereço do linkedin é obrigatório";
+        //public const string PostalCode = "O cep é obrigatório";
+        //public const string PostalCodeInvalid = "O formato do cep está inválido";
+        //public const string Linkedin = "O seu endereço do linkedin é obrigatório";
         #endregion
 
         public UserValidator()
@@ -29,16 +29,18 @@ namespace ShareBook.Domain.Validators
 
             RuleFor(u => u.Password)
               .NotEmpty()
-              .WithMessage(Password);
+              .WithMessage(Password)
+              .Must(x => x != null && x.Length >= 6 && x.Length <= 32)
+              .WithMessage("Senha deve ter entre 6 e 32 letras.");
                 
         }
 
-        private bool PostalCodeIsValid(string postalCode)
-        {
-            Regex Rgx = new Regex(@"^\d{5}-\d{3}$");
-            if (string.IsNullOrEmpty(postalCode) || !Rgx.IsMatch(postalCode)) return false;
+        //private bool PostalCodeIsValid(string postalCode)
+        //{
+        //    Regex Rgx = new Regex(@"^\d{5}-\d{3}$");
+        //    if (string.IsNullOrEmpty(postalCode) || !Rgx.IsMatch(postalCode)) return false;
 
-            return true;
-        }
+        //    return true;
+        //}
     }
 }
