@@ -1,26 +1,23 @@
 ﻿// <auto-generated />
+using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Migrations;
-using Microsoft.EntityFrameworkCore.Storage;
-using Microsoft.EntityFrameworkCore.Storage.Internal;
-using ShareBook.Domain.Enums;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using ShareBook.Repository;
-using System;
 
-namespace ShareBook.Repository.Migrations
+namespace ShareBook.Infra.Data.Migrations
 {
     [DbContext(typeof(ApplicationDbContext))]
-    [Migration("20190127011841_FacilitatorNotes")]
-    partial class FacilitatorNotes
+    [Migration("20191221005846_Initial")]
+    partial class Initial
     {
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "2.0.3-rtm-10026")
-                .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
+                .HasAnnotation("ProductVersion", "2.2.6-servicing-10079")
+                .HasAnnotation("Relational:MaxIdentifierLength", 64);
 
             modelBuilder.Entity("ShareBook.Domain.Address", b =>
                 {
@@ -28,21 +25,21 @@ namespace ShareBook.Repository.Migrations
                         .ValueGeneratedOnAdd();
 
                     b.Property<string>("City")
-                        .HasColumnType("varchar(30)")
+                        .HasColumnType("varchar(50)")
                         .HasMaxLength(30);
 
                     b.Property<string>("Complement")
-                        .HasColumnType("varchar(30)")
+                        .HasColumnType("varchar(50)")
                         .HasMaxLength(30);
 
                     b.Property<string>("Country")
-                        .HasColumnType("varchar(30)")
+                        .HasColumnType("varchar(50)")
                         .HasMaxLength(30);
 
                     b.Property<DateTime?>("CreationDate");
 
                     b.Property<string>("Neighborhood")
-                        .HasColumnType("varchar(30)")
+                        .HasColumnType("varchar(50)")
                         .HasMaxLength(30);
 
                     b.Property<string>("Number")
@@ -58,7 +55,7 @@ namespace ShareBook.Repository.Migrations
                         .HasMaxLength(30);
 
                     b.Property<string>("Street")
-                        .HasColumnType("varchar(50)")
+                        .HasColumnType("varchar(80)")
                         .HasMaxLength(50);
 
                     b.Property<Guid>("UserId");
@@ -142,6 +139,10 @@ namespace ShareBook.Repository.Migrations
 
                     b.Property<DateTime?>("CreationDate");
 
+                    b.Property<string>("NickName")
+                        .HasColumnType("varchar(64)")
+                        .HasMaxLength(64);
+
                     b.Property<string>("Note")
                         .HasColumnType("varchar(2000)")
                         .HasMaxLength(2000);
@@ -186,7 +187,7 @@ namespace ShareBook.Repository.Migrations
                     b.Property<DateTime?>("CreationDate");
 
                     b.Property<string>("Details")
-                        .HasColumnType("varchar(max)");
+                        .HasColumnType("varchar(1000)");
 
                     b.Property<bool>("IsSuccess");
 
@@ -235,6 +236,14 @@ namespace ShareBook.Repository.Migrations
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd();
 
+                    b.Property<bool>("Active")
+                        .ValueGeneratedOnAdd()
+                        .HasDefaultValueSql("1");
+
+                    b.Property<bool>("AllowSendingEmail")
+                        .ValueGeneratedOnAdd()
+                        .HasDefaultValue(true);
+
                     b.Property<DateTime?>("CreationDate");
 
                     b.Property<string>("Email")
@@ -247,7 +256,9 @@ namespace ShareBook.Repository.Migrations
                         .HasMaxLength(200);
 
                     b.Property<DateTime>("HashCodePasswordExpiryDate")
-                        .HasColumnType("datetime2(7)");
+                        .HasColumnType("datetime(6)");
+
+                    b.Property<DateTime>("LastLogin");
 
                     b.Property<string>("Linkedin")
                         .HasColumnType("varchar(100)")
@@ -295,7 +306,7 @@ namespace ShareBook.Repository.Migrations
                         .OnDelete(DeleteBehavior.Cascade);
 
                     b.HasOne("ShareBook.Domain.User", "User")
-                        .WithMany()
+                        .WithMany("BooksDonated")
                         .HasForeignKey("UserId");
 
                     b.HasOne("ShareBook.Domain.User", "UserFacilitator")
